feat: streamline analysis page by removing loading state and error handling from home component
